+++ src/main/java/org/traccar/protocol/TotemProtocolEncoder.java
@@ -1,12 +1,12 @@
 /*
  * Copyright 2015 Irving Gonzalez
- * Copyright 2016 - 2019 Anton Tananaev (anton@traccar.org)
+ * Copyright 2015 - 2019 Anton Tananaev (anton@traccar.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,27 +16,49 @@
  */
 package org.traccar.protocol;
 
+import org.traccar.StringProtocolEncoder;
 import org.traccar.model.Command;
 import org.traccar.Protocol;
 import org.traccar.helper.Checksum;
 
-public class TotemProtocolEncoder extends TotemProtocolSmsEncoder {
+public class TotemProtocolEncoder extends StringProtocolEncoder {
 
     public TotemProtocolEncoder(Protocol protocol) {
         super(protocol);
     }
 
-    private String encodeCommand(String commandString) {
-        String builtCommand = String.format("$$%04dCF%s", 10 + commandString.getBytes().length, commandString);
-        return String.format("%s%02X", builtCommand, Checksum.xor(builtCommand));
+    protected String getCommandString(Command command) {
+        switch (command.getType()) {
+            case Command.TYPE_CUSTOM:
+                return formatCommand(command, "%s,%s", Command.KEY_DEVICE_PASSWORD, Command.KEY_DATA);
+            case Command.TYPE_REBOOT_DEVICE:
+                return formatCommand(command, "%s,006", Command.KEY_DEVICE_PASSWORD);
+            case Command.TYPE_FACTORY_RESET:
+                return formatCommand(command, "%s,007", Command.KEY_DEVICE_PASSWORD);
+            case Command.TYPE_GET_VERSION:
+                return formatCommand(command, "%s,056", Command.KEY_DEVICE_PASSWORD);
+            case Command.TYPE_POSITION_SINGLE:
+                return formatCommand(command, "%s,012", Command.KEY_DEVICE_PASSWORD);
+            // Assuming PIN 8 (Output C) is the power wire, like manual says but it can be PIN 5,7,8
+            case Command.TYPE_ENGINE_STOP:
+                return formatCommand(command, "%s,025,C,1", Command.KEY_DEVICE_PASSWORD);
+            case Command.TYPE_ENGINE_RESUME:
+                return formatCommand(command, "%s,025,C,0", Command.KEY_DEVICE_PASSWORD);
+            default:
+                return null;
+        }
     }
 
     @Override
-    protected String encodeCommand(Command command) {
+    protected Object encodeCommand(Command command) {
 
         initDevicePassword(command, "000000");
 
-        return encodeCommand(super.getCommandString(command));
+        String commandString = getCommandString(command);
+        String builtCommand = String.format("$$%04dCF%s", 10 + commandString.getBytes().length, commandString);
+
+        return String.format("%s%02X", builtCommand, Checksum.xor(builtCommand));
+
     }
 
 }
